refactor: Wave 2 G — introduce Phantom Type IDs (MessageID/ContactID/ReportID/UserID)
Define phantom-typed primary keys in internal/ids and re-export via store/ to
prevent silently mixing IDs across domain entities. Conversions to int64 are
confined to the sqlc boundary; all other layers use phantom types.

Coverage: store/ services/ handlers/ scanner/ channels/ ai/ cmd/verify/ + tests.

Bug catches surfaced & fixed by the type discipline:
- int↔int64 mismatches in 7 functions (HandleTaskCompletion, ProcessOnDemand-
  Translation, GetReportTranslations, GetUserAliases/AddUserAlias,
  LogAIInference, InvalidateTaskTranslation).

Boundary decisions:
- types.EnrichedMessage.SenderID and logger.DecisionLog.TaskID stay raw int64
  (those packages cannot import store without circular deps); enricher shims
  bridge UserID → int64.
- handlers.parsePathID returns int64; callers wrap with the appropriate
  phantom type since path IDs target either messages or reports.
- TenantID deferred — repo uses tenant_email string as the tenant key; the
  backlog memo's TenantID int64 was aspirational, not real.

Verified: go build ./... && go vet ./... && bash unit-test.sh && golangci-lint run ./...

+++ logger/structured_logger.go
@@ -12,7 +12,7 @@ type DecisionLog struct {
 	Source    string    `json:"source"`
 	Room      string    `json:"room"`
 	State     string    `json:"state"`
-	TaskID    *int      `json:"task_id,omitempty"`
+	TaskID    *int64    `json:"task_id,omitempty"`
 	Task      string    `json:"task"`
 	Reasoning string    `json:"reasoning,omitempty"`
 }
